Inline rename actions to ast crate
