+++ lang/ast/src/exp/local_comatch.rs
@@ -10,6 +10,7 @@ use printer::{
 
 use crate::{
     ctx::{values::TypeCtx, LevelCtx},
+    rename::{Rename, RenameCtx},
     ContainsMetaVars, HasSpan, HasType, Occurs, Shift, ShiftRange, Substitutable, Substitution,
     Zonk, ZonkError,
 };
@@ -146,5 +147,13 @@ impl ContainsMetaVars for LocalComatch {
             self;
 
         cases.contains_metavars() || inferred_type.contains_metavars()
+    }
+}
+
+impl Rename for LocalComatch {
+    fn rename_in_ctx(&mut self, ctx: &mut RenameCtx) {
+        self.ctx = None;
+        self.inferred_type = None;
+        self.cases.rename_in_ctx(ctx);
     }
 }
