+++ lang/transformations/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod lifting;
-pub mod renaming;
 pub mod xfunc;
 
 pub use lifting::*;
-pub use renaming::*;
 pub use xfunc::*;
