--- lang/transformations/src/renaming/mod.rs
@@ -1,6 +1,0 @@
-mod ast;
-mod ctx;
-mod util;
-
-pub use ast::Rename;
-pub use ctx::Ctx;
